Add separate files for each class
