+++ captitalize_decorator.rb
@@ -1,0 +1,6 @@
+require_relative 'decorator'
+class CapitalizeDecorator < Decorator
+  def correct_name
+    super.capitalize
+  end
+end
