+++ nameable.rb
@@ -3,28 +3,3 @@ class Nameable
     raise NotImplementedError
   end
 end
-
-class Decorator < Nameable
-  def initialize(nameable)
-    @nameable = nameable
-    super()
-  end
-
-  def correct_name
-    @nameable.correct_name
-  end
-end
-
-class CapitalizeDecorator < Decorator
-  def correct_name
-    super.capitalize
-  end
-end
-
-class TrimmerDecorator < Decorator
-  def correct_name
-    return super[0, 10] if super.length > 10
-
-    super
-  end
-end
