+++ student.rb
@@ -2,7 +2,7 @@ require_relative 'person'
 
 class Student < Person
   def initialize(age, classroom, name = 'unknown', parent_permission: true)
-    super(age, name, parent_permission: parent_permission)
+    super(age, name, parent_permission:)
     @classroom = classroom
   end
