+++ teacher.rb
@@ -2,7 +2,7 @@ require_relative 'person'
 
 class Teacher < Person
   def initialize(age, specialization, name = 'unknown', parent_permission: true)
-    super(age, name, parent_permission: parent_permission)
+    super(age, name, parent_permission:)
     @specialization = specialization
   end
